Fixed bad calculation of offset/filesize ofr grib.

+++ LibTkGL/TclData/tclGRIB.h
@@ -53,7 +53,7 @@ int        GRIB_FileClose(Tcl_Interp *Interp,char *Id);
 GRIB_File* GRIB_FileGet(Tcl_Interp *Interp,char *Id);
 int        GRIB_FilePut(Tcl_Interp *Interp,GRIB_File *File);
 
-int     GRIB_FieldRead(Tcl_Interp *Interp,char *Name,char *File,int Key);
+int     GRIB_FieldRead(Tcl_Interp *Interp,char *Name,char *File,long Key);
 void    GRIB_FieldFree(TData *Data);
 void    GRIB_FieldSet(TData *Data);
 int     GRIB_FieldList(Tcl_Interp *Interp,GRIB_File *File,int Mode,char *Var);
